Fix flaw instance not being up to date after sync
After syncing a flaw with Bugzilla upon creating it from the API, the
flaw instance was not updated and if the Jira sync was enabled, a save
within its task sync would override the data and the Bugzilla data would
be lost. In particular, the `meta_attr` field would not be updated, even
if the Bugzilla bug is actually created. This caused errors when trying
to update the data from the API.

With this change, after the Bugzilla sync, the instance is updated so
that we are working with the latest data, and other mixins don't clash
with it.

Closes OSIDB-2628.