Ignore SLA if update stream specifies it's not applicable
The rhsa_sla_applicable field in update streams specifies whether SLA
is applicable or not, and in case it's not, SLA policies should not be
applied to the tracker related to the update stream.
Before this commit, SLA policies were applied regardless of the value
of this field. With this commit, this field is properly used to
determine whether SLA policies apply or not.

Closes OSIDB-2612.