Make flaw comments independent of Bugzilla
This commit redesigns the FlawComment model as well as the API to make
them independent of Bugzilla sync. Before this commit, the comment info
was being stored in meta_attr, which is filled during the sync with
Bugzilla. Some info was also shared in the base model (text) but wasn't
being sent through the API, while other info was completely lost
(creator, is_private).

With this commit, the info is stored in the model, and during sync it is
updated. The info in the model is also serialized and sent via the API.