Allow standalone IncludeMetaAttrMixin Serializers
Currently when include_meta_attr wasn't provided neither from
request nor from context passed from the parent serializer usage
of any serializer which takes advantage of IncludeMetaAttrMixin
might resulted in an error when not used with prepopulated
include_meta_attr in the context. We should be able to use
IncludeMetaAttrMixin as a standalone serializers as well.