Fix changelog for is_triage removal
With PR #564, the `is_triage` property was removed. This change was
added in the 'Changed' section of the changelog.

However, after further consideration, this change should have been
considered technical since this property was never even properly
implemented or visible in the API, so it's not even going into the
'Removed' section and just outright removed from the changelog to keep
things tidy.