Do not allow empty strings in CVE ID
The field `cve_id` is set as null=True, blank=True and unique=True. One
issue with this is that it can have both `None` and empty string values,
but only one value with empty string is allowed due to uniqueness. What
we actually want, is for the values to either have a valid CVE, or be
null.

Django doesn't execute the validations if the field is blank=True and we
have an empty string. However we do not want to set blank=False since
this would not allow sending a null value either. In
https://code.djangoproject.com/ticket/4136 this issue was addressed, but
only for forms, not on the model itself, and we want to make sure that
the data is consistent no matter where we modify it from (API, shell,
etc).

To overcome this issue, this commit modifies the save method so that an
empty string value is properly converted to `None` before saving.

Closes OSIDB-2702.