Improve affect validation error messages
Some validation error messages in the Affect model have been improved to
include more info on which affect is raising the error. This is useful
when performing bulk operations on affects, since a validation error
will pop up with no info on which affect caused it, and the flaw may
have lots of affects.

Closes OSIDB-2893.