Add switch to turn collectors on/off
This commit implements a way to turn off each collector using an
environment variable. Turning it off means it won't be caught by celery,
and therefore it won't be run automatically, although the collectors can
still be run manually from a shell even if they have been turned off.

There is an environment variable for each of the collector groups:
Bugzilla, Jira, Errata Tool, NVD and OSV. This means they can each be
turned on/off individually.

The collectors are turned on by default.

Closes OSIDB-2884.