Allow creating Jira task for a flaw on demand
OSIDB-authored flaws should create a Jira task when created. But for old
flaws inherited from SFM2, although they should be considered
OSIDB-authored (they don't come from a collector), they don't have an
associated Jira task and right now there is no way to create and
associate a task for them. These flaws can be recognised as they should
have an empty workflow state.

This commit adds the option to force the creation of a Jira task if the
flaw does not have one. When updating a flaw through the API, setting
the query parameter `create_jira_task` will trigger the creation of the
Jira task if it does not exist, and set the `task_key` and workflow
state fields.

Closes OSIDB-2882.