Add BZSync manager
The BZSync manager is the sync manager responsible for handling the
bzsync asynchronously.

The bzsync is still able to work both synchronously or asynchronously,
depending on an env variable. When working synchronously, it will work
the same way as until now. When working asynchronously, the sync manager
will come into play, sending the async task to celery.

Closes OSIDB-3106.
Closes OSIDB-3107.