Log exceptions in exception handler
By handling the exception and returning the
response with the exception text, we are
effectively throwing away an usefull piece of
information which is then missing in the logs.
This way, we preserve both, nice informative
message for the user and exception with full
traceback in the logs for us.