Cannot modify CVE of existing flaw
Since commit 654f09ddd69769b1b479773a51678002fc3e7ef1, the serializer
order for flaws changed, now running the Jira serializer before the
Bugzilla sync one. This causes the flaw instance to be updated to the
database before the bzsync is run, but the bzsync relied on the old
instance not being updated yet, and therefore the bzsync changes that
required the old instance for comparison are not being applied correctly
anymore.

With this commit, instead of relying on the instance not being saved by
the time bzsync is run, we copy the old instance without saving it to
the db, and pass it to the bzsync for it to use as needed.

Closes OSIDB-3102.