Move SLA policies to ps_constants
Instead of having the SLA policies in a file in OSIDB's codebase, the
file has been moved to the ps_constants repo, and the ps_constants
collector in OSIDB now picks the SLA policies from there.

To facilitate the collection of the SLA policies into OSIDB, some SLA
related objects have been converted to Django models, so that the data
from the collector can be stored and updated as with other ps_constants
data.

Closes OSIDB-3220.