Alert when Bugzilla sync failed
Now that bzsync can be asynchronous, it can be the case that the flaw is
correctly saved, but the sync to Bugzilla failed for some reason. In
this scenario, there may be data discrepancies between OSIDB and
Bugzilla, and as of now, there is no way for the user to know this
happened.

With this commit, when there is an error during bzsync, an alert is
raised on the flaw, so that the user knows that there may be data not
synced to Bugzilla.

Closes OSIDB-3252.