Fix IntegrityError during Tracker filing
This happened occasionaly when a tracker was filed
already in BTS and before the filing process
finished in OSIDB, Tracker collector was scheduled
to already download such tracker and when the
filing process got to actually saving the tracker
in OSIDB, there was already a Tracker with same
external_system_id created by collector.

In such case we just skip the save as it already
happen.