Implement complex conditions in workflow state requirements
Before this commit, the state requirements in a workflow were a list of
checks all of which had to be satisfied before the state advanced to the
next step.

With this commit, now it is possible to use more complex conditions in
the requirements, combining AND and OR logical operators. By default,
the requirements still all use an AND if no logical condition is
specified, making it backwards compatible. But in the list of
requirements we can now pass a dict such as

```
{
  "condition": "OR",
  "requirements": [
    ...
  ]
}
```

which specifies a logical operator (AND/OR) and a list of requirements
on which the operator will be applied.

The conditions can be nested for more complex conditions. In the
previous example, one of the requirements inside the condition could be
another condition.