Sync external references to Jira tracker links
Before this commit, when filing Jira trackers, external/upstream
references were created as links in the Jira issue, but this was only
done during tracker creation, and no further changes in the references
on OSIDB side were reflected in already created trackers.

With this commit, changes to references on OSIDB will create/update
links in existing Jira trackers. Links are kept unique by their URL, so
if a reference with a given URL changes its description, it will change
the Jira link description as well, but if the URL changes, it will
create a new Jira link.

Closes OSIDB-3733.