Implement SLA exclusion policies
An SLA policy with a null SLA will be used to exclude trackers under certain
conditions from getting an SLA and therefore a due date. Exclusion SLAs will
take priority over other SLAs, so if a condition is met for a tracker to be
excluded, it will not get a due date even if it matches other SLAs.

To define an exclusion policy, one can set the sla to `null` in the
YAML, like so

```
name: Excluded components
description: Components which do not need SLA
conditions:
  affect:
    - PS component in:
      - kpatch
      - kmatch
      - kcatch
sla: null
```

This will exclude all trackers with affects that meet the condition from
having a due date.