Implement 'in' operator in SLA conditions
The 'in' operator allows to check if an attribute is in a list of
values, described in a YAML list. For example, one can check if the
`ps_component` is any of three values like

```
conditions:
  affect:
    - PS component in:
      - kpatch
      - kmatch
      - kcatch
```

Additionally, the 'not in' operator has also been implemented as the
negation of the 'in' operator.