Update resolution of Obsolete trackers to Not affected
With new VEX justifications being enabled for the "Not a bug" Jira
resolution, we need to be able to reserve a specific tracker resolution
for those trackers that were filed erroneously. Since no VEX justification
exists for such state, we need to decide on a specific resolution that
will be used for this purpose. Obsolete seems like a good fit.

Yes, the current usage of Obsolete is all over the place, but we can clean
that up going forward at least a bit and amend our docs to recommend it
as a resolution for invalid trackers. If a tracker was opened in error,
it is likely that the underlying component/product is not affected,
so we should display it as such on the CVE page as well.