Introduce tracker special handling field
Add the special handling field to the Tracker model, which will be
synced to the corresponding field in Jira issues of 'Vulnerability'
type. In Jira, this field is an array which can contain a set of certain
values, so the same has been done in OSIDB.

This field has also been exposed to the API as a read only field.

Closes OSIDB-4065.