OSIDB-4204: Record CISA CVSS scores from cve.org as CISA's
This commit changes the CVE.org collector to correctly label ADP CVSS
scores from CISA as being CISA's and not CVEORG which was previously the
case.

Additionally, it adjusts the way the collector works so that it is
possible to update CVSS scores **only**, as it previously would only
allow create operations.