OSIDB-4199: Handle tracking of NIST/CISA CVSS scores in Jira trackers
With this commit, JIRA engineering trackers with a RH CVSS score < 7.0
but for which there exist NIST/CISA CVSS scores >= 7.0 will now meet the
mod7 criteria and will be included in the CVSS field in the
aforementioned tracker.

This will help downstream engineering teams to know why a specific
tracker was created when the RH CVSS score does not meet the usual
criteria.