Introduce affects v2
This commit modifies the Affect model to use ps_update_stream instead of
ps_module as one of its core identifiers. This adds one level more of
granularity to the affects.

It also adds a migration that expands the current affects to v2 by using
their trackers, which also exist at the stream level, or if there are no
trackers, expanding them by using the active streams from the product
definitions.

Since this is a big structural change, a database view that mimics the
old affects has been created to be able to maintain backward
compatibility, although they will be read-only. The v1 affects are
calculated from the v2 affects, taking a candidate from them using the
following rules:
1. For a flaw/component, we group v2 affects by ps_module. This gives us
   the opposite of the expansion done in the data migration.
2. If there are NOTAFFECTED affects in the group, they are ignored,
   unless all are NOTAFFECTED.
3. From the remaining affects, we take the one with the highest impact.
4. If there is still a tie, we break the tie using the created_dt or
   ultimately the uuid.
5. The remaining candidate affect is used as the representative for the
   v1 affect.
6. The trackers and cvss scores from all the affects v2 in the group are
   aggregated in this v1 affect.