Denormalize cve_id from Flaw to Affects and Trackers
This commit mirrors the cve_id field from Flaw into Affects and
Trackers since it's a pretty common operation to need to reference the
CVE id to which a given affect/tracker belongs to.

In the case of Affects this is usually not a very complicated operation
as we only have to traverse a single one-to-many relationship, however
in the case of Trackers we have to traverse a many-to-many relationship
to Affect and then back to Flaw.

Both operations require at least one JOIN, with this mirroring it
requires zero joins and the performance burden lies on the write
operations. Given that CVE ids for a Flaw rarely change (close to
never) this is considered a good trade-off.

This change also opens the door to parallel requests from clients based
off of a CVE ID, meaning a Flaw's data, its affects and trackers can be
fetched in parallel, so the performance bonus can be felt downstream
too.