OSIDB-4280: Provide mechanism for migrating affect/tracker ps_component
This commit introduces a management command, "migrate_components" which
will migrate all affects and trackers for a given ps_module,
ps_component to a new ps_component.

This is done so that we can easily fix incorrect/non-existing/historical
components into the correct components in batches in the context of
OSIDB-4280.