OSIDB-4367: Relax PURL and ps_component mismatch validation
In the context of backfilling PURLs into OSIDB, it may happen that
ps_components extrapolated from PURLs may not match the originally
provided ps_component.

This could be due to an error in the assigned ps_component or in the
PURL parsing, however this should not be a blocker for such an automatic
or pseudo-automatic operation.

This relaxation also allows users to override the ps_component when the
PURL parsing does not work properly.

Closes OSIDB-4367