Fix unembargo and set_public_nested
The unembargo and set_pulic_nested methods rely on traversing the
related models via its related_objects property. This worked when
affects and trackers were in a m2m relationship, however since now an
affect has a ForeignKey to a single tracker, this does not work anymore
and the chain breaks, leaving the trackers unembargoed/internal. This
commit fixes this by also traversing forward relationships (like a
ForeignKey) and unembargoing or setting to public the relevant entities
if needed.