Adapt trackers module to affects v2
The trackers module has been adapted to the use of affects v2. The
changes in the API have been as minimal as possible to avoid breaking
changes, so the format of the API calls and responses is the same.

The biggest change is that before the 'modules_components' object was
implied to be unique for every module and affect, but with affects v2,
there can be many affects for a single module, for each of its streams.