Perform tracker download linking in a single task
Before the introduction of (a)sync managers, when a tracker was
collected the linking back to its affects was done sequentially, but now
it is done in two separate async tasks. In the first task the tracker is
left orphaned, and the linking task is scheduled, but until it is picked
up by celery and processed, the user may see that some trackers have
temporarily disappeared from the flaw/affect.

This commit fixes this by doing the process of downloading and linking
the tracker in one single task, removing the necessity for separate sync
managers for the linking process.

Closes OSIDB-4069.