Make PURLs mandatory for middleware products
A validation is added at the Affect model ensuring that new affects for
middleware products contain a PURL, and that existing affects for
middleware products do not remove their PURL if they already have one.

For existing affects for middleware products that do not have a PURL,
nothing is enforced, to not block from working with legacy affects.

Closes OSIDB-4646.