Prioritize affects with subpath in PURL in v1
When choosing a v2 affect as the representative for the v1 endpoint,
prioritize affects that have a subpath in its PURL over everything else.
If more than one affect has a PURL with subpath, or no affect has a PURL
with subpath, the previously existing criteria for choosing the
representative apply.

Closes OSIDB-4645.