Introduce PURLField for handling PURLs in Django models
As the usage of PURL increases within our project, it is clear that we
need a streamlined way of using them within the context of Django and
OSIDB.

By introducing this model field, it is no longer necessary to convert
PURL strings into PackageURLs, validate them or normalize them ad-hoc as
the field handles all of this automatically.

By simply setting a model's PURLField to either a string or object,
developers can be certain that their PURL is valid, normalized (e.g.
consistent order of qualifiers) and individual PURL attributes can be
accessed directly without the need to explicitly convert to a PackageURL
object.