Revert conditional affect uniqueness on component/purl
The conditional uniqueness on PURL was intended for the use of the purl
field to track binray files, but since in the end we will use the
subpackage_purls for that, we do not need that constraint anymore. The
affects can still be unique per component, and the subpackage_purls
array can contain as many PURLs as we want.