Convert sync managers to proxy models
None of the sync manager concrete models introduce any schema changes,
instead the only difference in implementation is Python-based, this
makes these models prime candidates for Proxy models, i.e. models that
share the same table but differ in python implementation.

The reasoning is so that it's easier to query and expose sync manager
data through e.g. a single API endpoint instead of many endpoints or a
very custom and complex endpoint.

This commit also fixes a nasty but easy to miss bug WRT the relationship
between sync managers and the respective entity that they sync: if a
sync manager row was to be deleted, the related Flaw/Tracker would be
deleted on cascade.