Mark validation methods explicitly instead of relying on magic
Before this commit, validation methods get automagically picked up by
the AlertMixin by browsing all attributes of the model class at runtime,
filtering by methods that start with `_validate_`, then executing all of
those methods that match the criteria.

This has a few downsides:
* It's not explicit and not immediately apparent, unless you go into the
  definition of AlertMixin nobody would know why these methods are
  getting picked up and executed, could be considered cargo cult.
* It has performance implications as the entire model instance's
  attributes are looped through every time an object is saved and/or
  validated.
* Using third-party models or even base Django models can lead to
  undefined behavior and methods which are not intended to be run can be
  ran as long as they meet the method naming criteria.

Instead, this commit introduces a new implementation which leverages a
`@validator` decorator which marks methods that are business logic
validations (explicit), the marked methods are then collected when
subclassing the ValidateMixin which means it only needs to be done once
per model (performant) and does not rely on method names (avoids UB).